fastapi_tutorials/requirements.txt
fastapi
uvicorn
python-jose
sqlalchemy
passlib
bcrypt
python-jose